adios vtx: fix windows warning conversion from size_t to int

+++ IO/ADIOS2/Testing/Cxx/TestIOADIOS2VTX_VTI3DRendering.cxx
@@ -77,7 +77,8 @@ int MPIGetSize()
 
 std::size_t TotalElements(const std::vector<std::size_t>& dimensions) noexcept
 {
-  return std::accumulate(dimensions.begin(), dimensions.end(), 1, std::multiplies<std::size_t>());
+  return std::accumulate(dimensions.begin(), dimensions.end(), static_cast<std::size_t>(1),
+    std::multiplies<std::size_t>());
 }
 
 void WriteBPFile3DVars(const std::string& fileName, const adios2::Dims& shape,
